+++ sdk/ai/Azure.AI.Projects/src/Custom/Inference/Inference.cs
@@ -7,7 +7,6 @@ using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
-using Azure.AI.Inference;
 using Azure.Core;
 using Azure.Core.Pipeline;
 using Azure.Identity;
@@ -21,109 +20,6 @@ namespace Azure.AI.Projects
     public partial class AIProjectClient : ClientConnectionProvider
     {
         #nullable enable
-        public ChatCompletionsClient GetChatCompletionsClient()
-        {
-            ChatCompletionsClientKey chatCompletionsClientKey = new();
-            ChatCompletionsClient chatClient = this.Subclients.GetClient(chatCompletionsClientKey, () => CreateChatCompletionsClient());
-            return chatClient;
-        }
-
-        private ChatCompletionsClient CreateChatCompletionsClient()
-        {
-            ClientConnection connection = this.GetConnection(typeof(ChatCompletionsClient).FullName!);
-
-            if (!connection.TryGetLocatorAsUri(out Uri? uri) || uri is null)
-            {
-                throw new InvalidOperationException("Invalid URI.");
-            }
-            uri = new Uri($"https://{uri.Host}/models");
-
-            AzureAIInferenceClientOptions options = new AzureAIInferenceClientOptions();
-            BearerTokenAuthenticationPolicy tokenPolicy = new BearerTokenAuthenticationPolicy((TokenCredential)connection.Credential!, AuthorizationScopes);
-            options.AddPolicy(tokenPolicy, HttpPipelinePosition.PerRetry);
-
-            return connection.CredentialKind switch
-            {
-                CredentialKind.ApiKeyString => new ChatCompletionsClient(uri, new AzureKeyCredential((string)connection.Credential!)),
-                CredentialKind.TokenCredential => new ChatCompletionsClient(uri, (TokenCredential)connection.Credential!, options),
-                _ => throw new InvalidOperationException($"Unsupported credential kind: {connection.CredentialKind}")
-            };
-        }
-
-        private record ChatCompletionsClientKey();
-
-        /// <summary>
-        /// Gets the embeddings client.
-        /// </summary>
-        /// <returns></returns>
-        public EmbeddingsClient GetEmbeddingsClient()
-        {
-            EmbeddingsClientKey embeddingsClientKey = new();
-            EmbeddingsClient embeddingsClient = this.Subclients.GetClient(embeddingsClientKey, () => CreateEmbeddingsClient());
-            return embeddingsClient;
-        }
-
-        private EmbeddingsClient CreateEmbeddingsClient()
-        {
-            ClientConnection connection = this.GetConnection(typeof(ChatCompletionsClient).FullName!);
-
-            if (!connection.TryGetLocatorAsUri(out Uri? uri) || uri is null)
-            {
-                throw new InvalidOperationException("Invalid URI.");
-            }
-            uri = new Uri($"https://{uri.Host}/models");
-
-            AzureAIInferenceClientOptions options = new AzureAIInferenceClientOptions();
-            BearerTokenAuthenticationPolicy tokenPolicy = new BearerTokenAuthenticationPolicy((TokenCredential)connection.Credential!, AuthorizationScopes);
-            options.AddPolicy(tokenPolicy, HttpPipelinePosition.PerRetry);
-
-            return connection.CredentialKind switch
-            {
-                CredentialKind.ApiKeyString => new EmbeddingsClient(uri, new AzureKeyCredential((string)connection.Credential!)),
-                CredentialKind.TokenCredential => new EmbeddingsClient(uri, (TokenCredential)connection.Credential!, options),
-                _ => throw new InvalidOperationException($"Unsupported credential kind: {connection.CredentialKind}")
-            };
-        }
-
-        private record EmbeddingsClientKey();
-
-        /// <summary>
-        /// Gets the image embeddings client.
-        /// </summary>
-        /// <returns></returns>
-        public ImageEmbeddingsClient GetImageEmbeddingsClient()
-        {
-            ImageEmbeddingsClientKey imageEmbeddingsClientKey = new();
-            ImageEmbeddingsClient imageEmbeddingsClient = this.Subclients.GetClient(imageEmbeddingsClientKey, () => CreateImageEmbeddingsClient());
-            return imageEmbeddingsClient;
-        }
-
-        private ImageEmbeddingsClient CreateImageEmbeddingsClient()
-        {
-            ClientConnection connection = this.GetConnection(typeof(ImageEmbeddingsClient).FullName!);
-
-            if (!connection.TryGetLocatorAsUri(out Uri? uri) || uri is null)
-            {
-                throw new InvalidOperationException("Invalid URI.");
-            }
-            uri = new Uri($"https://{uri.Host}/models");
-
-            AzureAIInferenceClientOptions options = new AzureAIInferenceClientOptions();
-            OverrideApiVersionPolicy overrideApiVersionPolicy = new OverrideApiVersionPolicy("2024-05-01-preview");
-
-            options.AddPolicy(overrideApiVersionPolicy, HttpPipelinePosition.PerRetry);
-            BearerTokenAuthenticationPolicy tokenPolicy = new BearerTokenAuthenticationPolicy((TokenCredential)connection.Credential!, AuthorizationScopes);
-            options.AddPolicy(tokenPolicy, HttpPipelinePosition.PerRetry);
-
-            return connection.CredentialKind switch
-            {
-                CredentialKind.ApiKeyString => new ImageEmbeddingsClient(uri, new AzureKeyCredential((string)connection.Credential!)),
-                CredentialKind.TokenCredential => new ImageEmbeddingsClient(uri, (TokenCredential)connection.Credential!, options),
-                _ => throw new InvalidOperationException($"Unsupported credential kind: {connection.CredentialKind}")
-            };
-        }
-
-        private record ImageEmbeddingsClientKey();
 
         /// <summary>
         /// Gets the OpenAI chat client.